requirements-dev.txt
attrs >=17.4
black >=22.12.0
bump-my-version
compas_invocations2
invoke >=0.14
ruff
sphinx_compas2_theme
twine
wheel
build
pyvis
pytest-randomly
fakeredis